+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file is used to list changes made in each version of MageTestStand.
 
+## 2.2.0 (Unreleased)
+- rewrite travis logic
+- add official sources for all versions
+
 ## 2.1.6
 - add Magento CE 1.9.2.4